Update container creation

+++ bsbdock.nexus_context/Nexus/NexusServer.py
@@ -79,7 +79,7 @@ NEXUS_SERVER_ADDRESS = ('', 4281)
 # 3600sec <> 12h ; After 12h expired distribution targets are removed
 NEXUS_SERVER_TIMEOUT = 3600
 # Re-announce after half the expiration time
-NEXUS_SERVER_LONGPOLL = NEXUS_SERVER_TIMEOUT / 2
+NEXUS_SERVER_LONGPOLL = int(NEXUS_SERVER_TIMEOUT / 2)
 # Delay of initial announcement of this server to the network
 INITIAL_ANNOUNCEMENT_DELAY = 5
